requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
WTForms==3.0.1
SQLAlchemy==2.0.23
Flask-Migrate==4.0.5
Pillow==10.1.0
pandas==2.1.2
numpy==1.26.1
matplotlib==3.8.1
scikit-learn==1.3.2
pymongo==4.6.0
python-dotenv==1.0.0
cryptography==41.0.5
Werkzeug==2.3.7
python-docx==0.8.11
markdown==3.5 
flask-cors==2.1.2